Added mask for inputs, ensure model can run on both cpu and gpu, added parts to train.py

+++ DataPrep.py
@@ -12,7 +12,7 @@ def tensorFromSequence(sequence):
     """
     Generate tensors from the sequence in numpy.
     """
-    output = Variable(torch.tensor(sequence))
+    output = Variable(torch.tensor(sequence)).long()
     return output
